move more components into subfolders, and fix some type issues
