+++ client/src/components/activities/NewActivity/NewActivity.tsx
@@ -1,7 +1,7 @@
 import modalIds from "@/lib/modal-ids";
 import N from "@/lib/theme/components/form.style";
 import DefaultInput from "@/lib/theme/components/input/DefaultInput.style";
-import TagSelector from "@components/TagSelector/TagSelector";
+import TagSelector from "@components/tags/TagSelector/TagSelector";
 import { Checkbox } from "@lib/theme/components/Checkbox";
 import { type NewActivity } from "@type/server/activity.types";
 import DateTimePicker from "./DateTimePicker";
