+++ client/src/components/tags/TagSelector/Filter.tsx
@@ -1,4 +1,4 @@
-import type { FilterProps } from "@/components/TagSelector/tag-selector.types";
+import type { FilterProps } from "@/components/tags/TagSelector/tag-selector.types";
 import FilterInput from "@/lib/theme/components/input/FilterInput.style";
 import { MdOutlineClear } from "react-icons/md";
 import S from "./TagSelector.style";
